refact: hide sidebar if 1 entry on homepage + rm provenance tab
