+++ src/js/components/SiteSider.tsx
@@ -1,5 +1,4 @@
-import type React from 'react';
-import { useCallback, useMemo } from 'react';
+import { useCallback } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 
 import type { MenuProps, SiderProps } from 'antd';
@@ -10,7 +9,7 @@ import { useSelectedScope } from '@/features/metadata/hooks';
 import { useNonFilterQueryParams, useSearchQuery } from '@/features/search/hooks';
 import { buildQueryParamsUrl } from '@/features/search/utils';
 import { useLanguage, useTranslationFn } from '@/hooks';
-import { useGetRouteTitleAndIcon, useIsInCatalogueMode, useNavigateToRoot } from '@/hooks/navigation';
+import { useIsInCatalogueMode, useNavigateToRoot } from '@/hooks/navigation';
 import type { MenuItem } from '@/types/navigation';
 import { BentoRoute, TOP_LEVEL_ONLY_ROUTES } from '@/types/routes';
 import { getCurrentPage } from '@/utils/router';
@@ -19,7 +18,15 @@ const { Sider } = Layout;
 
 type OnClick = MenuProps['onClick'];
 
-const SiteSider = ({ collapsed, setCollapsed }: { collapsed: boolean; setCollapsed: SiderProps['onCollapse'] }) => {
+const SiteSider = ({
+  collapsed,
+  setCollapsed,
+  items,
+}: {
+  collapsed: boolean;
+  setCollapsed: SiderProps['onCollapse'];
+  items: MenuItem[];
+}) => {
   const navigate = useNavigate();
   const location = useLocation();
   const language = useLanguage();
@@ -30,7 +37,7 @@ const SiteSider = ({ collapsed, setCollapsed }: { collapsed: boolean; setCollaps
   const currentPage = getCurrentPage(location);
 
   const navigateToRoot = useNavigateToRoot();
-  const { fixedProject, scope, scopeSet } = useSelectedScope();
+  const { scope, scopeSet } = useSelectedScope();
 
   const handleMenuClick: OnClick = useCallback(
     ({ key }: { key: string }) => {
@@ -55,37 +62,6 @@ const SiteSider = ({ collapsed, setCollapsed }: { collapsed: boolean; setCollaps
     },
     [navigate, filterQueryParams, otherQueryParams, location.pathname]
   );
-
-  const createMenuItem = useCallback(
-    (key: string, label: string, icon?: React.ReactNode, children?: MenuItem[]): MenuItem => ({
-      key,
-      icon,
-      children,
-      label: t(label),
-    }),
-    [t]
-  );
-
-  const getRouteTitleAndIcon = useGetRouteTitleAndIcon();
-
-  const menuItems: MenuItem[] = useMemo(() => {
-    const items = [createMenuItem(BentoRoute.Overview, ...getRouteTitleAndIcon(BentoRoute.Overview))];
-
-    if (scope.project) {
-      // Only show provenance if we're not at the top level, since the giant list of context-less datasets is confusing.
-      items.push(createMenuItem(BentoRoute.Provenance, ...getRouteTitleAndIcon(BentoRoute.Provenance)));
-    }
-
-    if (BentoRoute.Beacon) {
-      items.push(createMenuItem(BentoRoute.Beacon, ...getRouteTitleAndIcon(BentoRoute.Beacon)));
-    }
-
-    if (BentoRoute.BeaconNetwork && (!scope.project || (scope.project && fixedProject))) {
-      items.push(createMenuItem(BentoRoute.BeaconNetwork, ...getRouteTitleAndIcon(BentoRoute.BeaconNetwork)));
-    }
-
-    return items;
-  }, [getRouteTitleAndIcon, createMenuItem, scope, fixedProject]);
 
   return (
     <Sider
@@ -116,7 +92,7 @@ const SiteSider = ({ collapsed, setCollapsed }: { collapsed: boolean; setCollaps
       <Menu
         selectedKeys={[currentPage]}
         mode="inline"
-        items={menuItems}
+        items={items}
         onClick={handleMenuClick}
         style={{ border: 'none' }}
       />
